make libdmi and implement C API
