+++ types.h
@@ -42,6 +42,52 @@ static inline u64 U64(u32 low, u32 high)
 }
 #endif
 
+enum dmi_types{
+	BIOS,
+	SYSTEM,
+	BASEBOARD,
+	CHASSIS,
+	PROCESSOR,
+	MEMORY_CONTROLLER,
+	MEMORY_MODULE,
+	CACHE,
+	PORT_CONNECTOR,
+	SYSTEM_SLOTS,
+	ON_BOARD_DEVICES,
+	OEM_STRINGS,
+	SYSTEM_CONFIGURATION_OPTIONS,
+	BIOS_LANGUAGE,
+	GROUP_ASSOCIATIONS,
+	SYSTEM_EVENT_LOG,
+	PHYSICAL_MEMORY_ARRAY,
+	MEMORY_DEVICE,
+	MEMORY_ERROR_32_BIT,
+	MEMORY_ARRAY_MAPPED_ADDRESS,
+	MEMORY_DEVICE_MAPPED_ADDRESS,
+	BUILT_IN_POINTING_DEVICE,
+	PORTABLE_BATTERY,
+	SYSTEM_RESET,
+	HARDWARE_SECURITY,
+	SYSTEM_POWER_CONTROLS,
+	VOLTAGE_PROBE,
+	COOLING_DEVICE,
+	TEMPERATURE_PROBE,
+	ELECTRICAL_CURRENT_PROBE,
+	OUT_OF_BAND_REMOTE_ACCESS,
+	BOOT_INTEGRITY_SERVICES,
+	SYSTEM_BOOT,
+	MEMORY_ERROR_64_BIT,
+	MANAGEMENT_DEVICE,
+	MANAGEMENT_DEVICE_COMPONENT,
+	MANAGEMENT_DEVICE_THRESHOLD_DATA,
+	MEMORY_CHANNEL,
+	IPMI_DEVICE,
+	POWER_SUPPLY,
+	ADDITIONAL_INFORMATION,
+	ONBOARD_DEVICES_EXTENDED_INFORMATION,
+	MANAGEMENT_CONTROLLER_HOST_INTERFACE
+};
+
 /*
  * Per SMBIOS v2.8.0 and later, all structures assume a little-endian
  * ordering convention.
